# docfx/restapi/azuredevopskatsV1.swagger.json
{"swagger":"2.0","info":{"version":"1.0","title":"AzureDevOpsKats.Web","description":"AzureDevOpsKats.Web","termsOfService":"Shareware","contact":{"name":"Stuart Shay","email":"sshay@yahoo.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"paths":{"/api/v1/cats":{"get":{"tags":["Cats"],"summary":"Get List of Cats","operationId":"Get","consumes":[],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"Success"}},"deprecated":false},"post":{"tags":["Cats"],"summary":"Create Cat","operationId":"Post","consumes":["application/json"],"produces":[],"parameters":[{"name":"value","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/CatCreateModel"}}],"responses":{"201":{"description":"Success"},"422":{"description":"Client Error"}},"deprecated":false}},"/api/v1/cats/{id}":{"get":{"tags":["Cats"],"summary":"Get Cat","operationId":"Get","consumes":[],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["Cats"],"summary":"Update Cat Properties","operationId":"Put","consumes":["application/json"],"produces":[],"parameters":[{"name":"id","in":"path","description":"Cat Id","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/CatUpdateModel"}}],"responses":{"200":{"description":"Returns the updated cat"},"422":{"description":"Validation error"}},"deprecated":false},"delete":{"tags":["Cats"],"summary":"Delete Cat","operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"204":{"description":"Success"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"CatCreateModel":{"required":["name","description"],"type":"object","properties":{"name":{"maxLength":50,"type":"string"},"description":{"maxLength":250,"type":"string"},"bytes":{"format":"byte","type":"string"}}},"CatUpdateModel":{"required":["name","description"],"type":"object","properties":{"name":{"maxLength":50,"type":"string"},"description":{"maxLength":250,"type":"string"}}}}}